CM3D is now better than ever with support for moving objects around with arrows, like in your favourite 3D modelling program. The coll editor was also reworked to support multiple files, NRE and the Prop Tool now preview the models and there's a lil' bit of documentation about each class now.

+++ src/ctrmap/Workspace.java
@@ -23,6 +23,9 @@ import java.util.prefs.Preferences;
 import javax.swing.JOptionPane;
 import javax.swing.SwingWorker;
 
+/**
+ * The core of CTRMap filesystem access which seamlessly takes care of everything the editor wants from the game.
+ */
 public class Workspace {
 
 	Preferences prefs;
@@ -351,7 +354,7 @@ public class Workspace {
 	public File getWorkspaceFile(ArchiveType arc, int fileNum) {
 		File wsFile;
 		wsFile = new File(getExtractionDirectory(arc).getAbsolutePath() + "/" + fileNum);
-		if (!wsFile.exists()) {
+		if (!wsFile.exists() && getArchive(arc).entries.size() > fileNum) {
 			try {
 				OutputStream os = new FileOutputStream(wsFile);
 				byte[] b = getArchive(arc).getDecompressedEntry(fileNum);
